+++ Editor/Scripts/ServiceLocator/ServiceLocatorCreator.cs
@@ -1,20 +1,23 @@
 using UnityEditor;
+using UnityEditor.SceneManagement;
 using UnityEngine;
 
 namespace NgoUyenNguyen.Editor
 {
     public static class ServiceLocatorCreator
     {
-        [MenuItem("GameObject/ServiceLocator/Global Scope")]
+        [MenuItem("GameObject/Service Locator/Global Scope")]
         private static void AddGlobal()
         {
             var go = new GameObject(ServiceLocator.GlobalServiceLocatorName, typeof(GlobalServiceLocatorBootstrapper));
+            EditorSceneManager.MarkSceneDirty(go.scene);
         }
 
-        [MenuItem("GameObject/ServiceLocator/Scene Scope")]
+        [MenuItem("GameObject/Service Locator/Scene Scope")]
         private static void AddScene()
         {
             var go = new GameObject(ServiceLocator.SceneServiceLocatorName, typeof(SceneServiceLocatorBootstrapper));
+            EditorSceneManager.MarkSceneDirty(go.scene);
         }
     }
 }